requirements.txt
Django==5.2.5
python-dotenv==1.1.1
psycopg2==2.9.10
djangorestframework==3.16.0
drf-spectacular==0.28.0
django-filter==25.1